Use correct winid when placing location-list signs
When ':LDiagnosticsPlace' is called from a location-list window, we need
to use the window-ID of the window the location-list belongs to and not
the window-ID of the location-list window itself. Otherwise users need
to keep the location-list window open and call ':LDiagnosticsClear' or
':LDiagnosticsToggle' from the same window. Note that every time a
quickfix or location-list window is opened, a new window-ID is assigned
to that window.