Remove location-list diagnostics on WinClosed
When the errors from a location-list are highlighted and the window is
closed, the signs and text-properties are automatically removed now,
since when a window is closed its location-list is deleted as well. It
makes no sense to keep the diagnostics around when the location-list
is gone.

Until now, a user had to run ":LDiagnosticsClear!" and clear all
diagnostics placed for other location-lists as well when one window was
closed in which ":LDiagnosticsClear" was forgotten to be run before
closing the window.