Add experimental auto popup on CursorMoved
Automatically show popup when cursor is moved on a line containing an
error. Currently, the popup shows up at the cursor.

Commands:

    :DiagnosticsToggleHover
    :LDiagnosticsToggleHover

TODO: Place popup in top/bottom left/center/right corner of the main Vim
window, or current split.

TODO: Group quickfix items by buffers so that we don't have to search
through the entire quickfix list. Then check the quickfix entry
'changedtick', and if it changed since the last access, obtain the new
quickfix with getqflist() or getloclist() and regroup the items again.