Check end_col in end_lnum before prop_add()
Before adding a text-property with prop_add(), we need to check that
end_col doesn't exceed the buffer line, but this has to be done with the
buffer line end_lnum which can be different from lnum (think of
multi-line text-property).

Also add a sanity check for the case when a buffer line is empty.