Remove text-property types when clearing
When text-properties are removed, and then :undo is executed to go back
to an earlier buffer version, sometimes text-properties re-appear in the
buffer. This happens because text-properties are stored with the text in
memory and are therefore restored after :undo or :redo. Unfortunately,
there is no way to remove them "permanently", i.e. from the entire undo
history. There are several ways to get rid of the highlighting when
changing to an earlier buffer version:

1. Use listener_add() and call prop_remove() whenever a text-property
   re-appears.
2. Same as 1., but use TextChanged instead.
3. Use prop_type_delete()

Here we decided to use item 3. When a text-property type doesn't exist
anymore, all text-properties using this type are skipped and won't be
highlighted. Unfortunately, the text-properties will still remain in the
buffer.

See also discussion in vim's issue 11655.