Add text-properties for all loaded buffers
Instead of adding text-properties only for displayed buffers and then
after BufWinEnter is triggered, we add them for all loaded buffers and
then when BufReadPost is triggered.

Initially this is more work than adding text-properties only for buffers
that are displayed in a window (think of a list of loaded buffers that
will never be displayed in a window). But it will lead to less surprises
since buffers can be loaded with bufload() and edited, for example, with
deletebufline() without being displayed in a window. Having
text-properties attached to the buffer as soon as the buffer is loaded
will make sure that the text-properties will move correctly as lines are
appended or deleted from the buffer without being displayed in a window.