Remove signs in current buffer on BufUnload
Unlike text-properties, signs are not stored with the text in memory.
Therefore, they can exist even when the buffer is unloaded. Here we
unplace the signs in the current buffer when the buffer is unloaded to
mimic the same behavior that text-properties have.

Bonus: when the last buffer in a 'group' is unloaded, all text-property
types and autocommands are removed and the associated data for that
'group' is cleared, essentially resulting in as if :[L]DiagnosticsClear
has been executed.